U8 u16 api sync GitHub

+++ driver/src/dma/xaie_dma_aieml.h
@@ -58,15 +58,15 @@ AieRC _XAieMl_DmaGetChannelStatus(XAie_DevInst *DevInst, XAie_LocType Loc,
 AieRC _XAieMl_DmaCheckBdChValidity(u8 BdNum, u8 ChNum);
 AieRC _XAieMl_MemTileDmaCheckBdChValidity(u8 BdNum, u8 ChNum);
 AieRC _XAieMl_DmaUpdateBdLen(XAie_DevInst *DevInst, const XAie_DmaMod *DmaMod,
-		XAie_LocType Loc, u32 Len, u8 BdNum);
+		XAie_LocType Loc, u32 Len, u16 BdNum);
 AieRC _XAieMl_ShimDmaUpdateBdLen(XAie_DevInst *DevInst,
-		const XAie_DmaMod *DmaMod, XAie_LocType Loc, u32 Len, u8 BdNum);
+		const XAie_DmaMod *DmaMod, XAie_LocType Loc, u32 Len, u16 BdNum);
 AieRC _XAieMl_DmaUpdateBdAddr(XAie_DevInst *DevInst, const XAie_DmaMod *DmaMod,
-		XAie_LocType Loc, u64 Addr, u8 BdNum);
+		XAie_LocType Loc, u64 Addr, u16 BdNum);
 AieRC _XAieMl_ShimDmaUpdateBdAddr(XAie_DevInst *DevInst,
 		const XAie_DmaMod *DmaMod, XAie_LocType Loc, u64 Addr,
-		u8 BdNum);
-AieRC _XAieMl_DmaSetBdIteration(XAie_DmaDesc *DmaDesc, u32 StepSize, u8 Wrap,
+		u16 BdNum);
+AieRC _XAieMl_DmaSetBdIteration(XAie_DmaDesc *DmaDesc, u32 StepSize, u16 Wrap,
 		u8 IterCurr);
 
 #endif /* XAIE_DMA_AIEML_H */
